refactor(RichTextEditor/InputBar): file structure
