+++ src/script/components/InputBar/components/PastedFileControls/PastedFileControls.tsx
@@ -24,7 +24,7 @@ import {TabIndex} from '@wireapp/react-ui-kit/lib/types/enums';
 import * as Icon from 'Components/Icon';
 import {t} from 'Util/LocalizerUtil';
 
-import {Config} from '../../../Config';
+import {Config} from '../../../../Config';
 
 interface PastedFileControlsProps {
   pastedFile: File;
