+++ src/script/components/InputBar/components/RichTextEditor/components/FormatToolbar/FormatButton/FormatButton.styles.ts
