+++ src/script/components/InputBar/components/RichTextEditor/components/FormatToolbar/FormatToolbar.tsx
