+++ src/script/components/InputBar/components/RichTextEditor/components/FormatToolbar/common/isNodeHeading/isNodeHeading.test.ts
