+++ src/script/components/InputBar/components/RichTextEditor/components/SendMessageButton/SendMessageButton.tsx
