+++ src/script/components/InputBar/components/RichTextEditor/components/SendMessageButton/index.ts
