+++ src/script/components/InputBar/components/RichTextEditor/index.ts
