+++ src/script/components/InputBar/components/RichTextEditor/nodes/EmojiNode.ts
