+++ src/script/components/InputBar/components/RichTextEditor/nodes/Mention.tsx
