+++ src/script/components/InputBar/components/RichTextEditor/nodes/MentionNode.tsx
