+++ src/script/components/InputBar/components/RichTextEditor/plugins/AutoFocusPlugin/AutoFocusPlugin.tsx
