+++ src/script/components/InputBar/components/RichTextEditor/plugins/CodeHighlightPlugin/CodeHighlightPlugin.tsx
