+++ src/script/components/InputBar/components/RichTextEditor/plugins/EditedMessagePlugin/getMentionMarkdownTransformer/getMentionMarkdownTransformer.ts
@@ -19,7 +19,7 @@
 
 import {TextMatchTransformer} from '@lexical/markdown';
 
-import {$createMentionNode, $isMentionNode, MentionNode} from 'Components/RichTextEditor/nodes/MentionNode';
+import {$createMentionNode, $isMentionNode, MentionNode} from '../../../nodes/MentionNode';
 
 // Cutom transformer for handling mentions when converting markdown to editor format.
 // Based on https://github.com/facebook/lexical/blob/main/packages/lexical-markdown/src/MarkdownTransformers.ts#L489
