+++ src/script/components/InputBar/components/RichTextEditor/plugins/EmojiPickerPlugin/EmojiItem.styles.ts
