+++ src/script/components/InputBar/components/RichTextEditor/plugins/EmojiPickerPlugin/index.ts
