+++ src/script/components/InputBar/components/RichTextEditor/plugins/GlobalEventsPlugin/GlobalEventsPlugin.tsx
