+++ src/script/components/InputBar/components/RichTextEditor/plugins/HistoryPlugin/HistoryPlugin.tsx
