+++ src/script/components/InputBar/components/RichTextEditor/plugins/InlineEmojiReplacementPlugin/inlineReplacements.ts
