+++ src/script/components/InputBar/components/RichTextEditor/plugins/MentionsPlugin/index.ts
