+++ src/script/components/InputBar/components/RichTextEditor/plugins/SendPlugin/SendPlugin.tsx
