+++ src/script/components/InputBar/components/RichTextEditor/theme.ts
@@ -1,0 +1,84 @@
+/*
+ * Wire
+ * Copyright (C) 2025 Wire Swiss GmbH
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see http://www.gnu.org/licenses/.
+ *
+ */
+
+export const theme = {
+  ltr: 'ltr',
+  rtl: 'rtl',
+  placeholder: 'editor-placeholder',
+  paragraph: 'editor-paragraph',
+  mentions: {
+    '@': `at-mentions`, // use the trigger name as the key
+    '@Focused': 'focused-mentions', // add the "Focused" suffix to style the focused mention
+  },
+  text: {
+    bold: 'editor-bold',
+    italic: 'editor-italic',
+    underline: 'editor-underline',
+    strikethrough: 'editor-strikethrough',
+    code: 'editor-inline-code',
+  },
+  quote: 'editor-quote',
+  list: {
+    ul: 'editor-list editor-list-unordered',
+    ol: 'editor-list editor-list-ordered',
+    listitem: 'editor-list__item',
+    nested: {
+      listitem: 'editor-list__item--nested',
+    },
+    olDepth: ['editor-list-ordered--1', 'editor-list-ordered--2', 'editor-list-ordered--3'],
+  },
+  heading: {
+    h1: 'editor-heading editor-heading--1',
+    h2: 'editor-heading editor-heading--2',
+    h3: 'editor-heading editor-heading--3',
+  },
+  code: 'editor-code',
+  codeHighlight: {
+    atrule: 'editor-tokenAtrule',
+    attr: 'editor-tokenAttr',
+    boolean: 'editor-tokenBoolean',
+    builtin: 'editor-tokenBuiltin',
+    cdata: 'editor-tokenCdata',
+    char: 'editor-tokenChar',
+    class: 'editor-tokenClass',
+    'class-name': 'editor-tokenClassName',
+    comment: 'editor-tokenComment',
+    constant: 'editor-tokenConstant',
+    deleted: 'editor-tokenDeleted',
+    doctype: 'editor-tokenDoctype',
+    entity: 'editor-tokenEntity',
+    function: 'editor-tokenFunction',
+    important: 'editor-tokenImportant',
+    inserted: 'editor-tokenInserted',
+    keyword: 'editor-tokenKeyword',
+    namespace: 'editor-tokenNamespace',
+    number: 'editor-tokenNumber',
+    operator: 'editor-tokenOperator',
+    prolog: 'editor-tokenProlog',
+    property: 'editor-tokenProperty',
+    punctuation: 'editor-tokenPunctuation',
+    regex: 'editor-tokenRegex',
+    selector: 'editor-tokenSelector',
+    string: 'editor-tokenString',
+    symbol: 'editor-tokenSymbol',
+    tag: 'editor-tokenTag',
+    url: 'editor-tokenUrl',
+    variable: 'editor-tokenVariable',
+  },
+};
