+++ src/script/components/InputBar/components/RichTextEditor/utils/getDomRangeRect.ts
