+++ src/script/components/InputBar/components/RichTextEditor/utils/getSelectionInfo.ts
